🛡️ Sentry: [test coverage improvement] duke-loader edge cases
This commit implements a suite of test cases targeting previously uncovered error
paths within the `DirectoryLoader`, `JImageReader`, `ZipLoader`, and `ZipReader`
components of `duke-loader`.

Specific additions include:
- `DirectoryLoader`: Tests rejecting path traversals (e.g. `..`), root relative
  paths, missing components, and dot/colon separated variants, confirming
  `LoadError::NotFound` returns.
- `JImageReader`: Tests for OOB offset bounds (`LoadError::JImageFormat`), invalid
  decompression payload triggers (`LoadError::Decompress`), and non-existent
  resource requests (`LoadError::NotFound`).
- `ZipLoader` & `ZipReader`: Validates `BOOT-INF` path resolution miss handling,
  CD signature corruption, truncated entry properties, EOF extensions, and nested
  JAR `LoadError::NotFound` mapping.
