+++ Cargo.toml
@@ -27,6 +27,9 @@ regex = "1"
 pedantic = { level = "warn", priority = -1 }
 nursery = { level = "warn", priority = -1 }
 
+[workspace.lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tarpaulin_include)'] }
+
 [profile.profiling]
 inherits = "release"
 debug = 1
