+++ crates/duke-loader/src/jimage.rs
@@ -828,84 +828,6 @@ mod tests {
     }
 
     #[test]
-    fn jimage_find_class_not_found() {
-        let reader = make_reader(vec![0; 10], 0, 0);
-        let result = reader.find_class("does/not/Exist");
-        assert!(matches!(result, Err(LoadError::NotFound { .. })));
-    }
-
-    #[test]
-    fn jimage_read_resource_not_found() {
-        let reader = make_reader(vec![0; 10], 0, 0);
-        let result = reader.read_resource("/java.base/Missing");
-        assert!(matches!(result, Err(LoadError::NotFound { .. })));
-    }
-
-    #[test]
-    fn jimage_read_resource_offset_overflow() {
-        let mut index = HashMap::new();
-        index.insert(
-            "r".to_string(),
-            ResourceInfo {
-                offset: u64::MAX,
-                compressed: 0,
-                uncompressed: 10,
-            },
-        );
-        let reader = JImageReader {
-            data: vec![0; 100],
-            resource_count: 1,
-            data_offset: 100, // This plus u64::MAX as usize will overflow
-            index,
-        };
-        let result = reader.read_resource("r");
-        assert!(matches!(result, Err(LoadError::JImageFormat { .. })));
-    }
-
-    #[test]
-    fn jimage_read_resource_truncated() {
-        let mut index = HashMap::new();
-        index.insert(
-            "r".to_string(),
-            ResourceInfo {
-                offset: 0,
-                compressed: 0,
-                uncompressed: 1000,
-            },
-        );
-        let reader = JImageReader {
-            data: vec![0; 10],
-            resource_count: 1,
-            data_offset: 0,
-            index,
-        };
-        let result = reader.read_resource("r");
-        assert!(matches!(result, Err(LoadError::JImageFormat { .. })));
-    }
-
-    #[test]
-    fn jimage_read_resource_decompress_error() {
-        let mut index = HashMap::new();
-        index.insert(
-            "r".to_string(),
-            ResourceInfo {
-                offset: 0,
-                compressed: 10,
-                uncompressed: 100,
-            },
-        );
-        let reader = JImageReader {
-            // Invalid deflate data
-            data: vec![0xFF; 10],
-            resource_count: 1,
-            data_offset: 0,
-            index,
-        };
-        let result = reader.read_resource("r");
-        assert!(matches!(result, Err(LoadError::Decompress { .. })));
-    }
-
-    #[test]
     fn read_resource_rejects_capacity_overflow() {
         // Attack: Provide an uncompressed size of u64::MAX.
         // If unpatched, Vec::with_capacity(usize::MAX) will panic with a capacity overflow.
